basic TT conversion of dw-free homepage
This only makes 2 rendering changes: h1s converted to h2s,
and updated the documentation link (although it still points
to LJ, because our wiki doesn't have a comparable page).

Still kind of terrible in terms of content, but we really
only care about the version in nonfree, so whatever.